Rearrange some files to remove padding
