+++ config/splat.hd.dra.yaml
@@ -37,8 +37,8 @@ segments:
       - [0x510, data]
       - [0x1F18, .data, save_mgr]
       - [0x2178, data]
-      - [0x2CF4, .data, 5298C] # array of strings
-      - [0x2D1C, .data, 5298C] # HD exclusive array of strings
+      - [0x2CF4, .data, 5087C] # array of strings
+      - [0x2D1C, .data, 5087C] # HD exclusive array of strings
       - [0x2D88, data]
       - [0x2ED0, .data, 5D6C4]
       - [0x2EF8, data]
@@ -72,16 +72,16 @@ segments:
       - [0x3C204, .rodata, collider]
       - [0x3C404, .rodata, demo]
       - [0x3C488, .rodata, 5087C] # func_800F0940
-      - [0x3C4A8, .rodata, 52860]
-      - [0x3C4D0, .rodata, 5298C] # func_800F298C
-      - [0x3C548, .rodata, 5298C] # strings
-      - [0x3C658, .rodata, 5298C] # HD exclusive array of strings
-      - [0x3C6F8, .rodata, 5298C] # DrawMenuFamiliars
-      - [0x3C718, .rodata, 5298C] # DrawMenu
-      - [0x3C758, .rodata, 5298C] # func_800F9F40
-      - [0x3C788, .rodata, 5298C] # MenuHandleCursorInput
-      - [0x3C7A0, .rodata, 5298C] # HandleMenu str
-      - [0x3C7C8, .rodata, 5298C] # HandleMenu jtbl
+      - [0x3C4A8, .rodata, 5087C]
+      - [0x3C4D0, .rodata, 5087C] # func_800F298C
+      - [0x3C548, .rodata, 5087C] # strings
+      - [0x3C658, .rodata, 5087C] # HD exclusive array of strings
+      - [0x3C6F8, .rodata, 5087C] # DrawMenuFamiliars
+      - [0x3C718, .rodata, 5087C] # DrawMenu
+      - [0x3C758, .rodata, 5087C] # func_800F9F40
+      - [0x3C788, .rodata, 5087C] # MenuHandleCursorInput
+      - [0x3C7A0, .rodata, 5087C] # HandleMenu str
+      - [0x3C7C8, .rodata, 5087C] # HandleMenu jtbl
       - [0x3CC18, .rodata, 5D6C4]
       - [0x3CC60, rodata]
       - [0x3CC78, .rodata, 627C4] # jtbl_800DCC78
@@ -165,8 +165,6 @@ segments:
       - [0x4FAE0, c, collider]
       - [0x5027C, c, demo]
       - [0x50F00, c, 5087C]
-      - [0x52EE4, c, 52860]
-      - [0x53010, c, 5298C]
       - [0x5D678, c, 5D6C4]
       - [0x62654, c, 627C4]
       - [0x63D1C, c, 63ED4]
